+++ altapay/__init__.py
@@ -1,5 +1,5 @@
 __title__ = 'altapay'
-__version__ = '1.5.6'
+__version__ = '1.5.7'
 __author__ = 'Coolshop.com'
 __license__ = 'MIT'
 __github_url__ = 'https://github.com/coolshop-com/AltaPay'
@@ -11,6 +11,8 @@ __api_base_url__ = {
 from altapay.api import API  # NOQA
 from altapay.callback import Callback  # NOQA
 from altapay.chargeback import ChargebackEvent  # NOQA
+from altapay.credit_card_wallet_initiate_app_payment import \
+    CreditCardWalletInitiateAppPayment  # NOQA
 from altapay.funding import CustomReport, Funding, FundingList  # NOQA
 from altapay.invoice import Invoice  # NOQA
 from altapay.payment import Payment  # NOQA